- begin moving things into different files - add start of pooling implementation
