+++ Cargo.toml
@@ -7,9 +7,11 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+actix = "0.8"
 actix-web = "1.0.8"
-diesel = { version = "1.4.2", features = [ "sqlite" ] }
+diesel = { version = "1.4.2", features = [ "r2d2", "sqlite" ] }
 diesel_migrations = "1.4.0"
+futures = "0.1"
 # dotenv = "0.14.1"
 
 # Needed for sqlite.
